% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{add_aesthetic}
\alias{add_aesthetic}
\title{Add an aesthetic to a grob object class.}
\usage{
add_aesthetic(
  grob_object,
  aesthetic,
  group = c("cells", "column_names", "column_headings"),
  value = NULL
)
}
\arguments{
\item{grob_object}{The R6 grob object class outputted by \code{\link{grob_matrix}}.}

\item{aesthetic}{The aesthetic the user wishes to add.}

\item{group}{The group of the grob matrix object the user wants to add the 
aesthetic to. For matrices, the user can add an aesthetic to the cells, the
column names or the column headings.}

\item{value}{A single value the user wants to apply to the group of matrix 
elements for the given aesthetic.}
}
\value{
The R6 grob object class with its aesthetics properties altered.
}
\description{
Add an aesthetic to a grob object class.
}
\examples{

library(dplyr)
df = data.frame(var1 = c(5, 14, 6, 10), var2 = c(3, 30, 17, 7))
df \%>\%
  grob_matrix() \%>\% 
  add_aesthetic(aesthetic = 'text_color', group = 'cells', value = 'red') \%>\%
  grob_col(width = 100) \%>\%
  .$grob \%>\%
  plot()

}
