% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{add_column_headings}
\alias{add_column_headings}
\title{Add column headings onto a matrix. Intended to be used with
the group_elements aesthetic within \code{\link{grob_col}}.}
\usage{
add_column_headings(mat, headings, heading_cols)
}
\arguments{
\item{mat}{The matrix the column headings will be added onto.}

\item{headings}{The headings to be added onto the initial matrix, 
in a list with each heading a separate element. The list must have
the same amount of elements as the \code{heading_cols} parameter.}

\item{heading_cols}{Which column positions of the initial matrix the \code{headings} 
will be placed above, in a list with each heading's column positions a separate element. 
The list must have the same amount of elements as the \code{headings} parameter.}
}
\value{
The initial matrix with column headings inserted into the 
first row.
}
\description{
Add column headings onto a matrix. Intended to be used with
the group_elements aesthetic within \code{\link{grob_col}}.
}
