% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{add_row_headings}
\alias{add_row_headings}
\title{Add row headings onto a matrix. Intended to be used with
the \code{group_elements} aesthetic within \code{\link{grob_col}} and \code{\link{ga_list}}.}
\usage{
add_row_headings(mat, headings = list(), heading_rows = list())
}
\arguments{
\item{mat}{The matrix the column headings will be added onto.}

\item{headings}{The headings to be added onto the initial matrix, 
in a list with each heading a separate element. The list must have
the same amount of elements as the \code{heading_rows} parameter.}

\item{heading_rows}{Which row positions of the initial matrix the \code{headings} 
will be placed to the left of, in a list with each heading's row positions a separate element. 
The list must have the same amount of elements as the \code{headings} parameter.}
}
\value{
The initial matrix with row headings inserted into the 
first column.
}
\description{
Add row headings onto a matrix. Intended to be used with
the \code{group_elements} aesthetic within \code{\link{grob_col}} and \code{\link{ga_list}}.
}
