% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{alter_columns}
\alias{alter_columns}
\title{Alter the values at specific columns of a matrix.}
\usage{
alter_columns(mat, value, columns = NULL)
}
\arguments{
\item{mat}{The matrix the user wishes to alter cells of.}

\item{value}{The single value that will replace specific cells of the matrix.}

\item{columns}{The columns the user wishes to alter. Can be numeric column
positions, or the user can input:
\itemize{
\item 'odd' - Only alter odd numbered columns.
\item 'even' - Only alter even numbered columns.
\item 'first' - Only alter the first column.
\item 'last' - Only alter the last column.
}
Also, the user can provide the column name of the column they wish to alter.}
}
\value{
A matrix with the desired columns altered.
}
\description{
Alter the values at specific columns of a matrix.
}
