% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{alter_rows}
\alias{alter_rows}
\title{Alter the values at specific rows of a matrix.}
\usage{
alter_rows(mat, value, rows = NULL)
}
\arguments{
\item{mat}{The matrix the user wishes to alter cells of.}

\item{value}{The single value that will replace specific cells of the matrix.}

\item{rows}{The rows the user wishes to alter. Can be numeric row positions, or 
the user can also input:
\itemize{
\item 'odd' - Only alter odd numbered rows.
\item 'even' - Only alter even numbered rows.
\item 'first' - Only alter the first row.
\item 'last' - Only alter the last row.
}
Also, the user can provide the row name of the column they wish to alter.}
}
\value{
A matrix with the desired rows altered.
}
\description{
Alter the values at specific rows of a matrix.
}
\examples{
df = data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
mat = aes_matrix(df, 'white')
alter_rows(mat = mat, value = 'red', rows = c(1, 2))
}
