% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{column_names_to_row}
\alias{column_names_to_row}
\title{Take a data.frame/matrix and insert its column names as the 
first row of the resulting matrix.}
\usage{
column_names_to_row(df)
}
\arguments{
\item{df}{The data.frame/matrix.}
}
\value{
A matrix of the initial data.frame/matrix with its column
names as the first row.
}
\description{
Take a data.frame/matrix and insert its column names as the 
first row of the resulting matrix.
}
