% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to_image_grob}
\alias{convert_to_image_grob}
\title{Converts a raw .png file to a grob, with flexible aesthetics.}
\usage{
convert_to_image_grob(.image, aes_list, height = numeric(), width = numeric())
}
\arguments{
\item{.image}{The local path to the raw .png file.}

\item{aes_list}{The list outputted by \code{\link{ga_list}} which gives the image grob its aesthetics.}

\item{height}{A numeric value designating the total height of the matrix grob in mm.}

\item{width}{A numeric value designating the total width of the matrix grob in mm.}
}
\value{
A grob of the raw .png file.
}
\description{
Converts a raw .png file to a grob, with flexible aesthetics.
}
