% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to_matrix_grob}
\alias{convert_to_matrix_grob}
\title{Converts a data.frame/matrix to a grob, with flexible aesthetics.}
\usage{
convert_to_matrix_grob(
  .df,
  aes_list = list(),
  height = numeric(),
  width = numeric()
)
}
\arguments{
\item{.df}{The data.frame/matrix to be converted to a grob.}

\item{aes_list}{The list outputted by \code{\link{ga_list}} which gives the data.frame/matrix grob its aesthetics.}

\item{height}{A numeric value designating the total height of the matrix grob in mm.}

\item{width}{A numeric value designating the total width of the matrix grob in mm.}
}
\value{
A grob of \code{.df}, with the corresponding aesthetics.
}
\description{
Converts a data.frame/matrix to a grob, with flexible aesthetics.
}
