% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_col.R
\name{grob_col}
\alias{grob_col}
\title{The grob-column function where an object is converted a grob. Works within \code{\link{grob_row}} and \code{\link{grob_layout}}.}
\usage{
grob_col(..., p = 1, padding = 1, aes_list = ga_list(), border = F,
  border_aes_list = ga_list(), title = "",
  title_aes_list = ga_list(), title_p = 0.2, hjust = 0.5,
  vjust = 0.5)
}
\arguments{
\item{...}{Either the object to be converted to a grob, or a combination of grob row's which need to be converted to sub-grob's.}

\item{p}{The numeric proportion of the width given to the outer grob_row which should be given to the grob-column outputted by this function. Defaults to 1.}

\item{aes_list}{The list outputted by \code{\link{ga_list}}, which controls aesthetics object within the grob-column.}

\item{border}{A TRUE/FALSE argument corresponding to whether or not a border around the outputted grob-column is desired. Defaults to FALSE.}

\item{border_aes_list}{The list outputted by \code{\link{ga_list}}, which controls aesthetics of the borders. Only two aesthetics that can be tweaked for borders are \code{border_color}, \code{border_width} and \code{border_sides}. Ignored if \code{border} is set to FALSE.}

\item{title}{A character string which will be displayed as the title of the grob-column.}

\item{title_aes_list}{The list outputted by \code{\link{ga_list}}, which controls aesthetics of the title of the grob-column.}

\item{title_p}{The numeric proportion of height within the grob-column and its allotted space which will be used by the title grob. A numeric value between 0 and 0.5.}

\item{hjust}{A numeric value between 0 and 1 which will determine the alignment of the grob horizontally within its designated area. A value of 0 means moving the grob all the way to the left, a value of 1 means moving the grob all the way to the right and a value of 0.5 means keeping the grob in the middle. Defaults to 0.5.}

\item{vjust}{A numeric value between 0 and 1 which will determine the alignment of the grob vertically within its designated area. A value of 0 means moving the grob all the way to the bottom, a value of 1 means moving the grob all the way to the top and a value of 0.5 means keeping the grob in the middle. Defaults to 0.5.}
}
\value{
An R6 class object which contains all the information needed to create the grob-column. The grob-column is obtained with \code{grob_col$grob}.
}
\description{
The grob-column function where an object is converted a grob. Works within \code{\link{grob_row}} and \code{\link{grob_layout}}.
}
