% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_matrix.R
\name{grob_matrix}
\alias{grob_matrix}
\title{Converts a data.frame/matrix to a grob, with flexible aesthetics. Used within
the GrobblR::convert_to_grob() function.}
\usage{
grob_matrix(df, gm_list, tot_height = numeric(), tot_width = numeric())
}
\arguments{
\item{df}{The data.frame/matrix to be converted to a grob.}

\item{gm_list}{A list which contains aesthetic parameters for the matrix grob.}

\item{tot_height}{A numeric value designating the total height of the matrix grob in mm.}

\item{tot_width}{A numeric value designating the total width of the matrix grob in mm.}
}
\value{
A grob of df, with the corresponding aesthetics.
}
\description{
Converts a data.frame/matrix to a grob, with flexible aesthetics. Used within
the GrobblR::convert_to_grob() function.
}
