% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_object_functions.R
\name{grob_text}
\alias{grob_text}
\title{Initialize a grob text object, to be used within \code{\link{grob_col}}.}
\usage{
grob_text(x)
}
\arguments{
\item{x}{A single character string.}
}
\value{
An R6 object of the grob matrix class.
}
\description{
Initialize a grob text object, to be used within \code{\link{grob_col}}.
}
\examples{

}
