% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_to_pdf.R
\name{grob_to_pdf}
\alias{grob_to_pdf}
\title{Converts a single grob to a pdf, or combines multiple grobs into a multiple page PDF document. Saves to the working directory. Uses the returned lists form grob_layout(), in order to properly size each page of the PDF document.}
\usage{
grob_to_pdf(..., file_name = character())
}
\arguments{
\item{...}{The single result from grob_layout() or series of results from grob_layout() which will be converted to a PDF document.}

\item{file_name}{The desired file name of the resulting PDF document in character format.}
}
\value{
A PDF document of the grob(s) which will be saved to the working directory.
}
\description{
Converts a single grob to a pdf, or combines multiple grobs into a multiple page PDF document. Saves to the working directory. Uses the returned lists form grob_layout(), in order to properly size each page of the PDF document.
}
