% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_grob}
\alias{view_grob}
\title{View a grob with a given width and height.}
\usage{
view_grob(grob, height = 100, width = 100, units = c("mm", "cm", "inches"))
}
\arguments{
\item{grob}{An object outputted by one of the following functions:
\itemize{

\item \code{\link{grob_matrix}}
\item \code{\link{grob_image}}
\item \code{\link{grob_row}}
\item \code{\link{grob_col}}
\item \code{\link{grob_layout}}

}}

\item{height}{The numeric height the user wishes to view the grob in.}

\item{width}{The numeric width the user wishes to view the grob in.}

\item{units}{The units the provided height and width are in.}
}
\description{
View a grob with a given width and height.
}
\examples{

data.frame(
  v2 = c(15, 4, 16, 11),
  v2 = c(10, 30, 3, 10)
  ) \%>\%
  grob_matrix() \%>\%
  view_grob()

}
