% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{aes_matrix}
\alias{aes_matrix}
\title{Create a matrix based off the dimensions of a data.frame/matrix and a single 
value to make up its cells.}
\usage{
aes_matrix(df, value)
}
\arguments{
\item{df}{A data.frame the resulting matrix will get its dimensions from.}

\item{value}{The single value that will make up the cells of the resulting matrix.}
}
\value{
A matrix based on the dimensions of \code{df} and \code{value}.
}
\description{
Create a matrix based off the dimensions of a data.frame/matrix and a single 
value to make up its cells.
}
\examples{
df = data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
aes_matrix(df, 'white')
}
