% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{alter_at}
\alias{alter_at}
\title{Flexibly alter the aesthetic of a grob matrix object at specific points of
the data.frame/matrix.}
\usage{
alter_at(
  grob_object,
  .f = NULL,
  ...,
  columns = NULL,
  rows = NULL,
  data = NULL,
  aesthetic = NULL,
  group = NULL
)
}
\arguments{
\item{grob_object}{The R6 grob object class outputted by \code{\link{grob_matrix}}.}

\item{.f}{A quosure style lambda \code{~ fun(.)}, which the user wants to apply
to the specific subset of cells.}

\item{...}{Logical predicates defined in terms of the variables in the initial 
data.frame/matrix, or if the user provides a new data.frame to evaluate via 
\code{data}. Multiple conditions are combined with &. Only rows where the condition 
evaluates to TRUE are evaluated. 

If no logical predicates provided, then the entire columns will be altered.}

\item{columns}{A character vector of column names of the initial data.frame/matrix
the user wishes to alter.}

\item{data}{A separate data.frame/matrix of the same dimensions as the initial 
data.frame/matrix which the \code{.f} function will be applied to. Any potential
filters will always be applied to the initial data.frame/matrix, regardless if 
a separate data.frame/matrix is supplied.}

\item{aesthetic}{Which aesthetic the user wants to make alterations to. If left 
\code{NULL}, the function will look for the most previous altered aesthetic, either
via \code{\link{add_aesthetic}}, or a previous application of \code{alter_at}.}

\item{group}{Which group of elements the user wants to make alterations to. If left 
\code{NULL}, the function will look for the most previous altered group, either
via \code{\link{add_aesthetic}}, or a previous application of \code{alter_at}.}
}
\value{
The R6 grob object class with its aesthetics properties altered.
}
\description{
Flexibly alter the aesthetic of a grob matrix object at specific points of
the data.frame/matrix.
}
\examples{

df = data.frame(var1 = c(5, 14, 6, 10), var2 = c(3, 30, 17, 7))
df \%>\%
  grob_matrix() \%>\%
  add_aesthetic(aesthetic = 'text_color', group = 'cells', value = 'red') \%>\%
  alter_at(
    .f = ~ 'blue',
    abs(var2 - var1) > 1
    ) \%>\%
  view_grob()

test_function = function(x) ifelse(x > 15, 2, 1)

df \%>\%
  grob_matrix() \%>\%
  alter_at(
    .f = ~ test_function(.),
    aesthetic = 'font_face',
    group = 'cells'
    ) \%>\%
  view_grob()

}
