% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{alter_cells}
\alias{alter_cells}
\title{Alter the values at specific row-column combinations of a matrix.}
\usage{
alter_cells(mat, value, rows = NULL, columns = NULL)
}
\arguments{
\item{mat}{The matrix the user wishes to alter cells of.}

\item{value}{The single value that will replace specific cells of the matrix.}

\item{rows}{The rows the user wishes to alter. See \code{\link{alter_rows}}
for information on special inputs.}

\item{columns}{The columns the user wishes to alter. See \code{\link{alter_columns}}
for information on special inputs.}
}
\value{
A matrix with the desired cells altered.
}
\description{
Alter the values at specific row-column combinations of a matrix.
}
