% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{alter_cells}
\alias{alter_cells}
\title{Alter the values at specific row-column combinations of a matrix.}
\usage{
alter_cells(mat, value, row_numbers = NULL, column_numbers = NULL)
}
\arguments{
\item{mat}{The matrix the user wishes to alter cells of.}

\item{value}{The single value that will replace specific cells of the matrix.}

\item{row_numbers}{The row numbers the user wishes to alter.}

\item{column_numbers}{The row numbers the user wishes to alter.}
}
\value{
A matrix with the desired cells altered.
}
\description{
Alter the values at specific row-column combinations of a matrix.
}
\examples{
df = data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
mat = aes_matrix(df, 'white')
alter_cells(mat = mat, value = 'red', row_numbers = c(1,2), column_numbers = 1)
}
