% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_matrix.R
\name{alter_columns}
\alias{alter_columns}
\title{Alter the values at specific columns of a matrix.}
\usage{
alter_columns(mat, value, column_numbers = NULL)
}
\arguments{
\item{mat}{The matrix the user wishes to alter cells of.}

\item{value}{The single value that will replace specific cells of the matrix.}

\item{column_numbers}{The column numbers the user wishes to alter.}
}
\value{
A matrix with the desired columns altered.
}
\description{
Alter the values at specific columns of a matrix.
}
\examples{
df = data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
mat = aes_matrix(df, 'white')
alter_columns(mat = mat, value = 'red', column_numbers = 1)
}
