% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to_matrix_grob}
\alias{convert_to_matrix_grob}
\title{Converts a data.frame/matrix to a grob, with flexible aesthetics.}
\usage{
convert_to_matrix_grob(
  df,
  aes_list = ga_list(),
  m_type = 1,
  height = numeric(),
  width = numeric(),
  padding = numeric(),
  units = c("mm"),
  text_cex_adj = 0.2
)
}
\arguments{
\item{df}{The data.frame/matrix to be converted to a grob.}

\item{aes_list}{The list outputted by \code{\link{ga_list}} which gives the data.frame/matrix grob its aesthetics.}

\item{m_type}{A integer value which indicates what the default aesthetics of the table will be. Default is 1. The possible options:
\enumerate{
\item Plain theme.
\item Table theme.
\item Column name theme.
\item Caption theme.
\item Title theme.
}}

\item{height}{A numeric value designating the total height of the matrix grob in mm.}

\item{width}{A numeric value designating the total width of the matrix grob in mm.}

\item{text_cex_adj}{A numeric value used to adjust the automatic text cex sizing.}
}
\value{
A grob of df, with the corresponding aesthetics.
}
\description{
Converts a data.frame/matrix to a grob, with flexible aesthetics.
}
