% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_list.R
\name{ga_list}
\alias{ga_list}
\title{Grob Aesthetic List}
\usage{
ga_list(
  aspect_ratio_multiplier = NULL,
  background_color = NULL,
  background_alpha = NULL,
  border_color = NULL,
  border_sides = NULL,
  border_width = NULL,
  font_face = NULL,
  group_elements = NULL,
  text_color = NULL,
  text_align = NULL,
  text_v_align = NULL,
  text_just = NULL,
  text_v_just = NULL,
  text_cex = NULL,
  text_font = NULL,
  text_rot = NULL,
  round_rect_radius = NULL,
  column_widths = NULL,
  column_widths_p = NULL,
  padding_p = NULL,
  cell_font_face = NULL,
  cell_group_elements = NULL,
  cell_background_color = NULL,
  cell_background_alpha = NULL,
  cell_border_color = NULL,
  cell_border_sides = NULL,
  cell_border_width = NULL,
  cell_text_color = NULL,
  cell_text_align = NULL,
  cell_text_v_align = NULL,
  cell_text_just = NULL,
  cell_text_v_just = NULL,
  cell_text_cex = NULL,
  cell_text_font = NULL,
  cell_text_rot = NULL,
  cell_round_rect_radius = NULL,
  cell_column_widths = NULL,
  cell_column_widths_p = NULL,
  cell_padding_p = NULL,
  colname_font_face = NULL,
  colname_group_elements = NULL,
  colname_background_color = NULL,
  colname_background_alpha = NULL,
  colname_border_color = NULL,
  colname_border_sides = NULL,
  colname_border_width = NULL,
  colname_text_color = NULL,
  colname_text_align = NULL,
  colname_text_v_align = NULL,
  colname_text_just = NULL,
  colname_text_v_just = NULL,
  colname_text_cex = NULL,
  colname_text_font = NULL,
  colname_text_rot = NULL,
  colname_round_rect_radius = NULL,
  colname_column_widths = NULL,
  colname_column_widths_p = NULL,
  colname_padding_p = NULL,
  maintain_aspect_ratio = NULL,
  n_lines = NULL,
  str_sep = NULL
)
}
\arguments{
\item{aspect_ratio_multiplier}{A numeric value which controls how much to increase/decrease the aspect ratio of images or ggplot objects.}

\item{background_color}{Controls the background color of the elements of the matrix. If the matrix has no rownames or colnames, the default is white. If the matrix has column names, the default is white-gray90 on every odd-even row. Used with matrices.}

\item{background_alpha}{Controls the background alpha/opacity of the elements of the matrix. Values are used in \code{grid::gpar()}. Default is 1.0. Used with matrices.}

\item{border_color}{Controls the color of the selected borders. Default is gray40. Used with matrices.}

\item{border_sides}{Controls the borders of the elements of the matrix. The input is a string with the possible words "top", "bottom", "left", "right" separated by commas. For example, "top, left, right" will put borders on the top, left and right side of the grid cell, but not the bottom. Default is "", or no borders. Used with matrices.}

\item{border_width}{Controls the line width density/thickness of the selected borders. Values are used in \code{grid::gpar()}. Default is 4. Used with matrices.}

\item{font_face}{Controls the font face of the elements of the matrix (i.e. bold, italic, etc.). Values are used in \code{grid::gpar()}. Default for table elements is normal, or 1. Default for column name elements is "bold", or 2. Used with matrices or character strings.}

\item{group_elements}{A TRUE/FALSE argument on whether like, adjacent matrix elements should be grouped together into a single element. Default is FALSE.}

\item{text_color}{Controls the text color of the elements of the matrix. Default for table elements and row names is black, and a gray-blue color for column names. Used with matrices or character strings.}

\item{text_align}{Controls where the text in each grid cell will be centered around, horizontally. A numeric value between 0 and 1, with 0 being all the way to the left of the grid cell, and 1 being all the way to the right of the grid cell. Default is 0.5. Can also input 'left', 'right' or 'center', which will also make edits to \code{text_just} to make the text completely left-justified, right-justified or centered, respectively. Used with matrices or character strings.}

\item{text_v_align}{Controls where the text in each grid cell will be centered around, vertically. A numeric value between 0 and 1, with 0 being all the way to the bottom of the grid cell, and 1 being all the way to the top of the grid cell. Default is 0.5. Can also input 'top', 'bottom' or 'center', which will also make edits to \code{text_v_just} to make the text completely top-justified, bottom-justified or centered, respectively. Used with matrices or character strings.}

\item{text_just}{Controls the horizontal justification of the text in the matrix. A numeric value between 0 and 1, with 0 being left justification and 1 being right justification. Default is 0.5, or center justification. Can also input 'left', 'right' or 'center', which will also make edits to \code{text_align} to make the text completely left-justified, right-justified or centered, respectively. Used with matrices or character strings.}

\item{text_v_just}{Controls the vertical justification of the text in the matrix. A numeric value between 0 and 1, with 0 being bottom justification and 1 being top justification. Default is 0.5, or center justification. Can also input 'top', 'bottom' or 'center', which will also make edits to \code{text_v_align} to make the text completely top-justified, bottom-justified or centered, respectively. Used with matrices or character strings.}

\item{text_cex}{Controls the size of the text within the matrix. Default is automatic text sizing based on the length of the elements within the matrix, the row heights and the column widths. Used with matrices or character strings.}

\item{text_font}{Controls the font family of the text within the matrix. Default is sans. Used with matrices or character strings.}

\item{text_rot}{Controls the rotation in degrees of the text within the matrix. Default is 0 degrees. Used with matrices or character strings. \strong{Please be aware that the automatic text sizing will not react properly if the text is angled at anything other than 0 degrees.}}

\item{round_rect_radius}{Controls the radius of the corners of the rectangles matrix text is laid on top of. Used with matrices.}

\item{column_widths}{If automatic column widths are not desired, the user can provide a vector of widths for each column in the matrix in whatever units are specified in the grob-layout. Used with matrices.}

\item{column_widths_p}{If automatic column widths are not desired, the user can provide a vector of width proportions (ideally adding to 1) for each column in the matrix in whatever units are specified in the grob-layout. Used with matrices. Overridden by \code{column_widths} argument.}

\item{padding_p}{Controls the amount of proportional padding around each matrix cell. Used with matrices.}

\item{cell_font_face}{Controls the \code{font_face} of matrix cells. Overridden by the \code{font_face} parameter.}

\item{cell_group_elements}{Controls the \code{group_elements} of matrix cells. Overridden by the \code{group_elements} parameter.}

\item{cell_background_color}{Controls the \code{background_color} of matrix cells. Overridden by the \code{background_color} parameter.}

\item{cell_background_alpha}{Controls the \code{background_alpha} of matrix cells. Overridden by the \code{background_alpha} parameter.}

\item{cell_border_color}{Controls the \code{border_color} of matrix cells. Overridden by the \code{border_color} parameter.}

\item{cell_border_sides}{Controls the \code{border_sides} of matrix cells. Overridden by the \code{border_sides} parameter.}

\item{cell_border_width}{Controls the \code{border_width} of matrix cells. Overridden by the \code{border_width} parameter.}

\item{cell_text_color}{Controls the \code{text_color} of matrix cells. Overridden by the \code{text_color} parameter.}

\item{cell_text_align}{Controls the \code{text_align} of matrix cells. Overridden by the \code{text_align} parameter.}

\item{cell_text_v_align}{Controls the \code{text_v_align} of matrix cells. Overridden by the \code{text_v_align} parameter.}

\item{cell_text_just}{Controls the \code{text_just} of matrix cells. Overridden by the \code{text_just} parameter.}

\item{cell_text_v_just}{Controls the \code{text_v_just} of matrix cells. Overridden by the \code{text_v_just} parameter.}

\item{cell_text_cex}{Controls the \code{text_cex} of matrix cells. Overridden by the \code{text_cex} parameter.}

\item{cell_text_font}{Controls the \code{text_font} of matrix cells. Overridden by the \code{text_font} parameter.}

\item{cell_text_rot}{Controls the \code{text_rot} of matrix cells. Overridden by the \code{text_rot} parameter.}

\item{cell_round_rect_radius}{Controls the \code{round_rect_radius} of matrix cells. Overridden by the \code{round_rect_radius} parameter.}

\item{cell_column_widths}{Controls the \code{column_widths} of matrix cells. Overridden by the \code{column_widths} parameter.}

\item{cell_column_widths_p}{Controls the \code{column_widths_p} of matrix cells. Overridden by the \code{column_widths_p} parameter.}

\item{cell_padding_p}{Controls the \code{padding_p} of matrix cells. Overridden by the \code{padding_p} parameter.}

\item{colname_font_face}{Controls the font face of column names. Overridden by the \code{font_face} parameter.}

\item{colname_group_elements}{Controls the \code{group_elements} of column names. Overridden by the \code{group_elements} parameter.}

\item{colname_background_color}{Controls the \code{background_color} of column names. Overridden by the \code{background_color} parameter.}

\item{colname_background_alpha}{Controls the \code{background_alpha} of column names. Overridden by the \code{background_alpha} parameter.}

\item{colname_border_color}{Controls the \code{border_color} of column names. Overridden by the \code{border_color} parameter.}

\item{colname_border_sides}{Controls the \code{border_sides} of column names. Overridden by the \code{border_sides} parameter.}

\item{colname_border_width}{Controls the \code{border_width} of column names. Overridden by the \code{border_width} parameter.}

\item{colname_text_color}{Controls the \code{text_color} of column names. Overridden by the \code{text_color} parameter.}

\item{colname_text_align}{Controls the \code{text_align} of column names. Overridden by the \code{text_align} parameter.}

\item{colname_text_v_align}{Controls the \code{text_v_align} of column names. Overridden by the \code{text_v_align} parameter.}

\item{colname_text_just}{Controls the \code{text_just} of column names. Overridden by the \code{text_just} parameter.}

\item{colname_text_v_just}{Controls the \code{text_v_just} of column names. Overridden by the \code{text_v_just} parameter.}

\item{colname_text_cex}{Controls the \code{text_cex} of column names. Overridden by the \code{text_cex} parameter.}

\item{colname_text_font}{Controls the \code{text_font} of column names. Overridden by the \code{text_font} parameter.}

\item{colname_text_rot}{Controls the \code{text_rot} of column names. Overridden by the \code{text_rot} parameter.}

\item{colname_round_rect_radius}{Controls the \code{round_rect_radius} of column names. Overridden by the \code{round_rect_radius} parameter.}

\item{colname_column_widths}{Controls the \code{column_widths} of column names. Overridden by the \code{column_widths} parameter.}

\item{colname_column_widths_p}{Controls the \code{column_widths_p} of column names. Overridden by the \code{column_widths_p} parameter.}

\item{colname_padding_p}{Controls the \code{padding_p} of column names. Overridden by the \code{padding_p} parameter.}

\item{maintain_aspect_ratio}{A TRUE/FALSE value which indicates whether the aspect ratio of the image should be maintained. Default is FALSE - meaning the image will be stretched to fit the designated grid area. Used with images.}

\item{n_lines}{The maximum number of lines is desired for the character string to be broken up into. Used with character strings.}
}
\value{
A list with all possible aesthetic elements that can be adjusted, with the class of "grob_aes_list".
}
\description{
Grob aesthetic list used to control aesthetics within grobblR.
}
