% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_object_functions.R
\name{grob_image}
\alias{grob_image}
\title{Grob Image}
\usage{
grob_image(x)
}
\arguments{
\item{x}{Either a ggplot object or a file path to png image.}
}
\value{
An R6 object of the grob image class.
}
\description{
Initialize a grob image object, to be used within \code{\link{grob_col}}.
}
\examples{


}
