% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_image.R
\name{grob_image}
\alias{grob_image}
\title{Converts a raw .png file to a grob, with flexible aesthetics.}
\usage{
grob_image(img_path, aes_list, height = numeric(), width = numeric())
}
\arguments{
\item{img_path}{The local path to the raw .png file.}

\item{aes_list}{The list outputted by \code{\link{ga_list}} which gives the image grob its aesthetics.}

\item{height}{A numeric value designating the total height of the matrix grob in mm.}

\item{width}{A numeric value designating the total width of the matrix grob in mm.}
}
\value{
A grob of the raw .png file.
}
\description{
Converts a raw .png file to a grob, with flexible aesthetics.
}
