% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_layout.R
\name{grob_layout}
\alias{grob_layout}
\title{The main grobblR function which contains and organizes the GrobblR::grob_col() and grobblR::grob_row() functions, giving the combined grob its shape.}
\usage{
grob_layout(..., height = 280, width = 216, padding = 5,
  grob_padding = 2, row_heights = c())
}
\arguments{
\item{...}{The combination of grob_row's and grob_col's which will help give the main grob outputted its shape and look.}

\item{height}{The numeric height of the grob in mm. Default is 280 mm - which is the height of an upright 8.5 x 11 in piece of copy paper.}

\item{width}{The numeric width of the grob in mm. Default is 216 mm - which is the width of an upright 8.5 x 11 in piece of copy paper.}

\item{padding}{The numeric amount of padding around the edge of the grob  in mm. Default is 5 mm.}

\item{grob_padding}{The numeric amount of padding for each individual grob within their designated grid of area in mm. Default is 2 mm.}

\item{row_heights}{If the user wants to designate specific row heights instead relying on the proportions within the grob_row functions, set this parameter equal to a vector of numeric values corresponding to the individual row heights in mm. Length must be equal to the number of grob_row function's on the upper most level of the grob.}
}
\value{
A list containing the total grob, and other values related to its formation - including total height/width of the grob, row heights and the paddings.
}
\description{
The main grobblR function which contains and organizes the GrobblR::grob_col() and grobblR::grob_row() functions, giving the combined grob its shape.
}
