% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_matrix.R
\name{grob_matrix}
\alias{grob_matrix}
\title{Converts a data.frame/matrix to a grob, with flexible aesthetics. Used within
the GrobblR::convert_to_grob() function.}
\usage{
grob_matrix(df, gm_obj, tot_height = numeric(), tot_width = numeric())
}
\arguments{
\item{df}{The data.frame/matrix to be converted to a grob.}

\item{gm_obj}{An object of class "grob_matrix" which contains aesthetic parameters for
the matrix grob. Created within the GrobblR::convert_to_grob() function.}

\item{tot_height}{A numeric value designating the total height of the matrix grob in mm.}

\item{tot_width}{A numeric value designating the total width of the matrix grob in mm.}
}
\value{
A matrix with each row representing a
}
\description{
Converts a data.frame/matrix to a grob, with flexible aesthetics. Used within
the GrobblR::convert_to_grob() function.
}
