% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_object_functions.R
\name{grob_matrix}
\alias{grob_matrix}
\title{Initialize a grob matrix object, to be used within \code{\link{grob_col}}.}
\usage{
grob_matrix(x)
}
\arguments{
\item{x}{Either a data.frame, a matrix or a vector.}
}
\value{
An R6 object of the grob matrix class.
}
\description{
Initialize a grob matrix object, to be used within \code{\link{grob_col}}.
}
\examples{

data.frame(
  v1 = c(15, 4, 16, 11),
  v2 = c(10, 30, 3, 10)
  ) \%>\%
  grob_matrix() \%>\%
  view_grob()

}
