% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_row.R
\name{grob_row}
\alias{grob_row}
\title{The grob row function which helps gives the grob from the GrobblR::grob_layout()
function its shape. Works hand in hand with the GrobblR::grob_col() function.}
\usage{
grob_row(..., p = 1, border = F, border_args = gpar())
}
\arguments{
\item{...}{A series of grob_col's.}

\item{p}{The numeric proportion of the given height which
should be given to sub_grob's outputted in this grob row. Defaults to 1.}

\item{border}{A TRUE/FALSE argument corresponding to whether or not a border
around the outputted sub-grob (or sub_grob's) is desired. Defaults to FALSE.}

\item{border_args}{A gpar() list which contain desired aesthetics for the border
around the outputted the sub-grob. Ignored if the border param is set to FALSE.}
}
\value{
An R6 class which contains all the information needed to carry on to
the grob_col's to create their sub-grob's.
}
\description{
The grob row function which helps gives the grob from the GrobblR::grob_layout()
function its shape. Works hand in hand with the GrobblR::grob_col() function.
}
