% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_to_pdf.R
\name{grob_to_pdf}
\alias{grob_to_pdf}
\title{Converts a single grob to a pdf, or combines multiple grobs into a multiple page PDF document. Saves to the working directory. Only takes grobs with heights and widths in units of mm.}
\usage{
grob_to_pdf(..., file_name = character(), add_page_numbers = F)
}
\arguments{
\item{...}{The single grob or series of grobs which will be converted to a PDF document.}

\item{file_name}{The desired file name of the resulting PDF document in character format.}
}
\value{
A PDF document of the grob(s) which will be saved to the working directory.
}
\description{
Converts a single grob to a pdf, or combines multiple grobs into a multiple page PDF document. Saves to the working directory. Only takes grobs with heights and widths in units of mm.
}
